Create inputString function
