+++ kiselev.sergey/P4/main.cpp
@@ -1,47 +1,16 @@
+#include <inputString.h>
 #include <iostream>
 #include "identicalChar.h"
-#include "resizeArray.h"
 int main()
 {
-  size_t memorySize = 3;
-  char* arr = static_cast<char*>(malloc(memorySize * sizeof(char)));
-  if (arr == nullptr)
+  char* string = kiselev::inputString();
+  if (!string)
   {
-    std::cerr << "Out of memory\n";
-    return 1;
+    std::cerr << "Failed string reading\n";
   }
-  size_t length = 0;
-  std::noskipws(std::cin);
-  if (!(std::cin >> arr[length]))
-  {
-    std::cerr << "Incorrect input\n";
-    free(arr);
-    return 1;
-  }
-  while (arr[length] != '\n')
-  {
-    length++;
-    if (length == memorySize)
-    {
-      size_t timeLength = memorySize;
-      memorySize += 5;
-      arr = kiselev::resizeArr(arr, memorySize, timeLength);
-      if (arr == nullptr)
-      {
-        std::cerr << "Out of memory\n";
-        return 1;
-      }
-    }
-    if (!(std::cin >> arr[length]))
-    {
-      break;
-    }
-  }
-  arr[length] = '\0';
-  std::skipws(std::cin);
-  const char secondArr[] = "abs";
-  size_t countIdentical = kiselev::identicalChar(arr, secondArr);
+  const char secondString[] = "abs";
+  size_t countIdentical = kiselev::identicalChar(string, secondString);
   std::cout << "Identical character: " << countIdentical << "\n";
-  free(arr);
+  free(string);
   return 0;
 }
