+++ kiselev.sergey/common/inputString.h
@@ -1,8 +1,9 @@
-#ifndef RESIZEARRAY_H
-#define RESIZEARRAY_H
+#ifndef INPUTSTRING_H
+#define INPUTSTRING_H
 #include <cstddef>
 namespace kiselev
 {
   char* resizeArr(char* array, size_t length, size_t length2);
+  char* inputString();
 }
 #endif
